Updating Hero component
